api/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coverage==6.0.2
cssselect2==0.4.1; python_version >= '3.6'
dj-rest-auth==2.1.11
django-appconf==1.0.5; python_version >= '3.6'
django-autoslug==1.9.8
django-environ==0.8.0
django-imagekit==4.0.2
django==3.2.8
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
easy-thumbnails==2.8rc1
iniconfig==1.1.1
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
packaging==21.0; python_version >= '3.6'
pilkit==2.0
pillow==8.4.0
pluggy==1.0.0; python_version >= '3.6'
psycopg2-binary==2.9.1
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.3.0; python_version >= '3.6'
pyparsing==3.0.0rc2; python_version >= '3.6'
pytest-django==4.4.0
pytest==6.2.5; python_version >= '3.6'
pytz==2021.3
pyyaml==6.0; python_version >= '3.6'
pyyml==0.0.2
reportlab==3.6.2; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
svglib==1.1.0; python_version >= '3'
tinycss2==1.1.0; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
uritemplate==4.1.1
webencodings==0.5.1
